# gh-118761: improve import time for ``base64``