--- a/Doc/whatsnew/3.14.rst
+++ b/Doc/whatsnew/3.14.rst
@@ -712,11 +712,9 @@ asyncio
 base64
 ------
 
-* Improve performance of :func:`base64.b16decode` by a factor ten by
-  removing an unnecessary regular expression. Consequently, :mod:`re`
-  is no more implicitly available as ``base64.re`` and import time of
-  :mod:`base64` is improved by a factor six.
-  (Contributed by Bénédikt Tran in :gh:`118761`.)
+* Improve the performance of :func:`base64.b16decode` by up to ten times,
+  and reduce the import time of :mod:`base64` by up to six times.
+  (Contributed by Bénédikt Tran, Chris Markiewicz, and Adam Turner in :gh:`118761`.)
 
 
 io