--- a/Lib/base64.py
+++ b/Lib/base64.py
@@ -4,7 +4,6 @@
 # Modified 30-Dec-2003 by Barry Warsaw to add full RFC 3548 support
 # Modified 22-May-2007 by Guido van Rossum to use bytes everywhere
 
-import re
 import struct
 import binascii
 
@@ -281,6 +280,8 @@ def b16decode(s, casefold=False):
     s is incorrectly padded or if there are non-alphabet characters present
     in the input.
     """
+    import re
+
     s = _bytes_from_decode_data(s)
     if casefold:
         s = s.upper()