--- a/Lib/base64.py
+++ b/Lib/base64.py
@@ -280,12 +280,10 @@ def b16decode(s, casefold=False):
     s is incorrectly padded or if there are non-alphabet characters present
     in the input.
     """
-    import re
-
     s = _bytes_from_decode_data(s)
     if casefold:
         s = s.upper()
-    if re.search(b'[^0-9A-F]', s):
+    if s.translate(None, delete=b'0123456789ABCDEF'):
         raise binascii.Error('Non-base16 digit found')
     return binascii.unhexlify(s)