--- a/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`base64` by a factor six. This is achieved by
+importing :mod:`re` on demand. Patch by Bénédikt Tran.