--- a/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
@@ -1,2 +1,4 @@
-Improve import time of :mod:`base64` by a factor six. This is achieved by
-importing :mod:`re` on demand. Patch by Bénédikt Tran.
+Improve performance of :func:`base64.b16decode` by a factor ten by
+removing an un-necessary regular expression. Consequently, :mod:`re`
+is no more implicitly available as ``base64.re`` and import time of
+:mod:`base64` is improved by a factor six. Patch by Bénédikt Tran.