--- a/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-13-06-54.gh-issue-118761.f8oADD.rst
@@ -1,4 +1,5 @@
-Improve performance of :func:`base64.b16decode` by a factor ten by
-removing an unnecessary regular expression. Consequently, :mod:`re`
-is no more implicitly available as ``base64.re`` and import time of
-:mod:`base64` is improved by a factor six. Patch by Bénédikt Tran.
+Improve the performance of :func:`base64.b16decode` by up to ten times
+by more efficiently checking the byte-string for hexadecimal digits.
+Reduce the import time of :mod:`base64` by up to six times,
+by no longer importing :mod:`re`.
+Patch by Bénédikt Tran, Chris Markiewicz, and Adam Turner.