@vstinner 2025-01-14
> @AA-Turner merged commit bbd3300 into python:main Jan 14, 2025

FYI @picnixz was recently promoted as a core dev and so can merge his own changes.