@chris-eibl 2025-01-14
> and reduce the import time of :mod:`base64` by up to six times.
9.5 ms down to  5.7 ms  is impressive but not six times - rather 60%?