@picnixz 2025-01-14
hyperfine benchmarks also take into account interpreter's startup and so, while `-X importtime` only takes into account the time needed to import the module itself (hence the differences).