@chris-eibl 2025-01-14
Sure, my bad. Importtime itself is of course six times faster.

Thank you all for the great work here in cPython!